+++ config-versions.lua
@@ -1,0 +1,3 @@
+testfiledir  = "testfiles-versions"
+checkpatterns = { "v%d+[%.%d]* %d+" }
+checkfiles = { "mypackage.sty" }