+++ CHANGELOG.md
@@ -9,6 +9,7 @@ this project uses date-based 'snapshot' version identifiers.
 ## [Unreleased]
 
 - Expand default value of `binaryfiles`
+- Variable `checkpatterns` (issue \#449)
 
 ## [2026-03-26]